final-pjt-back/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.0
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
